+++ test/LambdaLogger.spec.ts
@@ -4,22 +4,6 @@ import { LogLevel } from "../src/LogLevel";
 import TestUtility from "./lib/TestUtility";
 
 describe("LambdaLogger", () => {
-  it("should be able to output Critical logs", async () => {
-    const message = "hello";
-
-    const lambdaLogger = LambdaLoggerFactory.create(
-      LogLevel.DEBUG,
-      TestUtility.extractSlackTokenFromEnv(),
-      TestUtility.extractSlackChannelFromEnv()
-    );
-    const logOutput = await lambdaLogger.critical(message, true);
-
-    const expectedContext = `CRITICAL \n ${util.inspect(message, false, null)}`;
-
-    expect(logOutput.logLevel).toBe("CRITICAL");
-    expect(logOutput.context).toBe(expectedContext);
-  });
-
   it("should be able to output Error logs", async () => {
     const message = "hello";
